requirements.txt
python-decouple==3.3
selenium==3.141.0
urllib3==1.25.9
